rename java sources
