+++ .gitignore
@@ -92,3 +92,4 @@ bin/
 /benchmarks/src/main/java/org/ucfs/*
 /benchmarks/logs/
 /examples/src/main/java/java7/JavaLexer.java
+/examples/src/main/java/java8/JavaLexer.java
