+++ examples/src/main/java/java8/Java.x
@@ -1,4 +1,4 @@
-package java7;
+package java8;
 
 import java.io.*;
